Minor syntax to change real64 to rk and int32 to ik

+++ src/lib/domain/conserved_vars.f90
@@ -1,5 +1,5 @@
 module mod_conserved_vars
-  use iso_fortran_env, only: int32, real64
+  use iso_fortran_env, only: ik => int32, rk => real64
 
   implicit none
 
@@ -7,10 +7,10 @@ module mod_conserved_vars
   public :: conserved_vars_t
 
   type :: conserved_vars_t
-    real(real64), allocatable, dimension(:, :) :: density
-    real(real64), allocatable, dimension(:, :) :: x_velocity
-    real(real64), allocatable, dimension(:, :) :: y_velocity
-    real(real64), allocatable, dimension(:, :) :: pressure
+    real(rk), allocatable, dimension(:, :) :: density
+    real(rk), allocatable, dimension(:, :) :: x_velocity
+    real(rk), allocatable, dimension(:, :) :: y_velocity
+    real(rk), allocatable, dimension(:, :) :: pressure
   end type
 
 contains
