+++ src/lib/math_constants.f90
@@ -1,8 +1,8 @@
 module math_constants
 
-  use iso_fortran_env, only : int32, real64
+  use iso_fortran_env, only: ik => int32, rk => real64
 
   implicit none
-  
-  real(real64), parameter :: pi = 4.0_real64*tan(1.0_real64)
+
+  real(rk), parameter :: pi = 4.0_rk * tan(1.0_rk)
 end module math_constants
